+++ src/colors/syntax/js.colors.ts
@@ -30,6 +30,14 @@ const jsColors: SyntaxColors = (tokens) => {
       settings: tokens.class.library
     },
     {
+      name: 'Javascript: Template expression punctuation',
+      scope: [
+        'punctuation.definition.template-expression.begin.js',
+        'punctuation.definition.template-expression.end.js'
+      ],
+      settings: tokens.puntuaction.embedded
+    },
+    {
       name: 'Arrow function. =>',
       scope: 'storage.type.function.arrow.js',
       settings: tokens.javascript.arrowFunction